requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
langchain-core
mysql-connector-python
python-dotenv
